[ci]: delete namespace when branch is deleted
Summary:
When a branch is deleted (on merge on some other reason), the
corresponding preview or feature namespace in the cluster should also be
deleted. Except for the main branch, if it's ever deleted ( most likely
by mistake) the corresponding project namespace will remain in place,
unless deleted by the cluster-admin.