[fix]: restrict workflow job triggers to branch refs
Summary:
Changes on tags and other refs should not trigger the workflows, since
the refs are used as parameters in other steps ( such as creating
namespaces) that have a certain semantic convention. This commit fixes
that by ensuring the job run only if a push in a made on a branch and
not tags.